running pytest
