+++ pyproject.toml
@@ -11,8 +11,14 @@ python = "^3.9"
 numpy = "^1.24.1"
 pandas = "^1.5.3"
 jellyfish = "^0.9.0"
+pytest = "^7.2.1"
 
 [tool.poetry.dev-dependencies]
+
+[tool.poetry.group.dev.dependencies]
+pytest-cov = "^4.0.0"
+pytest = "^7.2.1"
+
 [tool.semantic_release]
 version_variable = "pyproject.toml:version" # version location
 branch = "main"                             # branch to make releases of
